Refactored packages and added scalastyle config
