+++ app/views/error_template.scala.html
@@ -14,8 +14,14 @@
  * limitations under the License.
  *@
 
-@()(implicit request: Request[_], messages: Messages)
+@(pageTitle: String, heading: String, message: String)(implicit request: Request[_], messages: Messages)
 
-@uk.gov.hmrc.managevatsubscriptionfrontend.views.html.main_template(title = "Hello from manage-vat-subscription-frontend", bodyClasses = None) {
-    <h1>Hello from manage-vat-subscription-frontend !</h1>
+@contentHeader = {
+  <h1>@heading</h1>
 }
+
+@mainContent = {
+  <p>@message</p>
+}
+
+@govuk_wrapper(appConfig = config.FrontendAppConfig, title = pageTitle, contentHeader = Some(contentHeader), mainContent = mainContent)
