+++ app/views/govuk_wrapper.scala.html
@@ -14,7 +14,7 @@
  * limitations under the License.
  *@
 
-@(appConfig: uk.gov.hmrc.managevatsubscriptionfrontend.AppConfig,
+@(appConfig: config.AppConfig,
   title: String,
   mainClass: Option[String] = None,
   mainDataAttributes: Option[Html] = None,
