+++ app/views/helloworld/hello_world.scala.html
@@ -14,14 +14,8 @@
  * limitations under the License.
  *@
 
-@(pageTitle: String, heading: String, message: String)(implicit request: Request[_], messages: Messages)
+@()(implicit request: Request[_], messages: Messages)
 
-@contentHeader = {
-  <h1>@heading</h1>
-}
-
-@mainContent = {
-  <p>@message</p>
-}
-
-@govuk_wrapper(appConfig = uk.gov.hmrc.managevatsubscriptionfrontend.FrontendAppConfig, title = pageTitle, contentHeader = Some(contentHeader), mainContent = mainContent)
+ @views.html.main_template(title = "Hello from manage-vat-subscription-frontend", bodyClasses = None) {
+  <h1>Hello from manage-vat-subscription-frontend !</h1>
+ }
