+++ app/views/main_template.scala.html
@@ -33,7 +33,7 @@
 }
 
 
-@govuk_wrapper(appConfig = uk.gov.hmrc.managevatsubscriptionfrontend.FrontendAppConfig,
+@govuk_wrapper(appConfig = config.FrontendAppConfig,
                title = title,
                mainClass = mainClass,
                bodyClasses = bodyClasses,
