+++ conf/application.conf
@@ -17,7 +17,7 @@ include "common.conf"
 appName="manage-vat-subscription-frontend"
 application.router=prod.Routes
 
-application.global=uk.gov.hmrc.managevatsubscriptionfrontend.FrontendGlobal
+application.global=config.FrontendGlobal
 
 play.filters.headers.contentSecurityPolicy= "default-src 'self' 'unsafe-inline' localhost:9000 localhost:9032 www.google-analytics.com data:"
